inter_exp/pyproject.toml
[project]
name = "inter-exp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fonttools>=4.57.0",
    "freetype-py>=2.5.1",
    "pillow>=11.2.1",
]
